Modified sequenceNum to int64_t

+++ include/depthai-shared/datatype/RawImgFrame.hpp
@@ -167,7 +167,7 @@ struct RawImgFrame : public RawBuffer {
     Specs fb;
     uint32_t category;     //
     uint32_t instanceNum;  // Which source created this frame (color, mono, ...)
-    int sequenceNum;       // increments for each frame
+    int64_t sequenceNum;   // increments for each frame
     Timestamp ts;          // generation timestamp, synced to host time
     Timestamp tsDevice;    // generation timestamp, direct device monotonic clock
